+++ tasks.py
@@ -6,6 +6,7 @@ import platform
 from pathlib import Path
 import base64
 from invoke import task
+import shutil
 
 _IS_WINDOWS = platform.system() == 'Windows'
 _PY_DEFAULT_VERSION = '3.9'
@@ -117,3 +118,35 @@ def test(c, report=False):
     c.run('pytest --cov ploomber ' + ('--cov-report html' if report else ''),
           pty=True)
     c.run('flake8')
+
+
+@task
+def install_git_hook(c, force=False):
+    """Installs pre-push git hook
+    """
+    path = Path('.git/hooks/pre-push')
+    hook_exists = path.is_file()
+
+    if hook_exists:
+        if force:
+            path.unlink()
+        else:
+            sys.exit('Error: pre-push hook already exists. '
+                     'Run: "invoke install-git-hook -f" to force overwrite.')
+
+    shutil.copy('.githooks/pre-push', '.git/hooks')
+    print(f'pre-push hook installed at {str(path)}')
+
+
+@task
+def uninstall_git_hook(c):
+    """Uninstalls pre-push git hook
+    """
+    path = Path('.git/hooks/pre-push')
+    hook_exists = path.is_file()
+
+    if hook_exists:
+        path.unlink()
+        print(f'Deleted {str(path)}.')
+    else:
+        print('Hook doesn\'t exist, nothing to delete.')