+++ tests/cli/test_examples.py
@@ -1,3 +1,4 @@
+import json
 from datetime import datetime, timedelta
 from unittest.mock import Mock
 from pathlib import Path
@@ -11,13 +12,61 @@ from ploomber.cli import cli
 
 
 def _mock_metadata(**kwargs):
-    default = dict(timestamp=datetime.now().timestamp(), branch='master')
+    default = dict(timestamp=datetime.now().timestamp(),
+                   branch=examples._DEFAULT_BRANCH)
     return {**default, **kwargs}
 
 
 @pytest.fixture(scope='session')
 def clone_examples():
-    examples.main(name=None, force=False)
+    examples.main(name=None, force=True)
+
+
+@pytest.mark.parametrize('argv, kwargs', [
+    [
+        ['examples'],
+        dict(name=None, force=False, branch=None, output=None),
+    ],
+    [
+        ['examples', '-n', 'name'],
+        dict(name='name', force=False, branch=None, output=None),
+    ],
+    [
+        ['examples', '--name', 'name'],
+        dict(name='name', force=False, branch=None, output=None),
+    ],
+    [
+        ['examples', '-f'],
+        dict(name=None, force=True, branch=None, output=None),
+    ],
+    [
+        ['examples', '--force'],
+        dict(name=None, force=True, branch=None, output=None),
+    ],
+    [
+        ['examples', '-b', 'some-branch'],
+        dict(name=None, force=False, branch='some-branch', output=None),
+    ],
+    [
+        ['examples', '--branch', 'some-branch'],
+        dict(name=None, force=False, branch='some-branch', output=None),
+    ],
+    [
+        ['examples', '--output', 'path/to/dir'],
+        dict(name=None, force=False, branch=None, output='path/to/dir'),
+    ],
+    [
+        ['examples', '-o', 'path/to/dir'],
+        dict(name=None, force=False, branch=None, output='path/to/dir'),
+    ],
+])
+def test_cli(monkeypatch, argv, kwargs):
+    mock = Mock()
+    monkeypatch.setattr(examples, 'main', mock)
+
+    CliRunner().invoke(cli.cli, argv, catch_exceptions=False)
+
+    mock.assert_called_once_with(**kwargs)
 
 
 def test_error_if_exception_during_execution(monkeypatch):
@@ -33,6 +82,19 @@ def test_error_if_exception_during_execution(monkeypatch):
         excinfo.value)
 
 
+def test_click_exception_isnt_shadowed_by_runtime_error(monkeypatch):
+    monkeypatch.setattr(
+        cli.cli_module.examples, 'main',
+        Mock(side_effect=ClickException('some click exception')))
+
+    runner = CliRunner()
+
+    result = runner.invoke(cli.cli, ['examples'])
+
+    assert result.exit_code == 1
+    assert result.output == 'Error: some click exception\n'
+
+
 def test_clones_in_home_directory(monkeypatch, tmp_directory):
     # patch home directory
     monkeypatch.setattr(examples, '_home', str(tmp_directory))
@@ -41,15 +103,14 @@ def test_clones_in_home_directory(monkeypatch, tmp_directory):
     mock_run = Mock()
     monkeypatch.setattr(examples.subprocess, 'run', mock_run)
 
-    # mock _list_example, otherwise this will fail since we aren't cloning the
-    # code
-    monkeypatch.setattr(examples, '_list_examples', lambda _: None)
+    # mock list, otherwise this will fail since we aren't cloning
+    monkeypatch.setattr(examples._ExamplesManager, 'list', lambda _: None)
 
     examples.main(name=None, force=False)
 
     # check clones inside home directory
     mock_run.assert_called_once_with([
-        'git', 'clone', '--depth', '1', '--branch', 'master',
+        'git', 'clone', '--depth', '1', '--branch', examples._DEFAULT_BRANCH,
         'https://github.com/ploomber/projects',
         str(Path(tmp_directory, 'projects'))
     ],
@@ -67,9 +128,8 @@ def test_change_default_branch(monkeypatch, tmp_directory):
     mock_run = Mock()
     monkeypatch.setattr(examples.subprocess, 'run', mock_run)
 
-    # mock _list_example, otherwise this will fail since we aren't cloning the
-    # code
-    monkeypatch.setattr(examples, '_list_examples', lambda _: None)
+    # mock list, otherwise this will fail since we aren't cloning
+    monkeypatch.setattr(examples._ExamplesManager, 'list', lambda _: None)
 
     examples.main(name=None, force=False, branch='custom-branch')
 
@@ -82,6 +142,29 @@ def test_change_default_branch(monkeypatch, tmp_directory):
                                      check=True)
 
 
+def test_does_not_download_again_if_no_explicit_branch_requested(
+        monkeypatch, tmp_directory):
+    dir_ = Path(tmp_directory, 'examples')
+    monkeypatch.setattr(examples, '_home', dir_)
+
+    examples.main(name=None, force=False)
+
+    # fake metadata to make it believe that we got if from another branch
+    meta = json.loads((dir_ / '.metadata').read_text())
+    meta['branch'] = 'some-other-branch'
+    (dir_ / '.metadata').write_text(json.dumps(meta))
+
+    # mock it so we test if we downloaded again
+    mock_run = Mock()
+    monkeypatch.setattr(examples.subprocess, 'run', mock_run)
+
+    # if called again but no force nor branch arg, it shouldn't download again
+    examples.main(name=None, force=False, branch=None)
+    examples.main(name='templates/ml-online', force=False, branch=None)
+
+    mock_run.assert_not_called()
+
+
 def test_home_default_value():
     assert examples._home == Path('~', '.ploomber')
 
@@ -90,9 +173,11 @@ def test_list(clone_examples, capsys):
     examples.main(name=None, force=False)
     captured = capsys.readouterr()
 
-    assert 'Basic' in captured.out
-    assert 'Intermediate' in captured.out
-    assert 'Advanced' in captured.out
+    assert examples._DEFAULT_BRANCH in captured.out
+    assert 'Ploomber examples' in captured.out
+    assert 'Templates' in captured.out
+    assert 'Guides' in captured.out
+    assert 'Cookbook' in captured.out
 
 
 def test_do_not_clone_if_recent(clone_examples, monkeypatch):
@@ -134,7 +219,7 @@ def test_clones_if_outdated(clone_examples, monkeypatch, capsys):
 
 
 def test_clones_if_different_branch(clone_examples, monkeypatch, capsys):
-    # mock metadata to make it look like another branch
+    # mock metadata to make it look like it's a copy from another branch
     metadata = _mock_metadata(branch='another-branch')
     monkeypatch.setattr(examples._ExamplesManager, 'load_metadata',
                         lambda _: metadata)
@@ -145,7 +230,7 @@ def test_clones_if_different_branch(clone_examples, monkeypatch, capsys):
     # prevent actual deletion
     monkeypatch.setattr(examples.shutil, 'rmtree', lambda _: None)
 
-    examples.main(name=None, force=False)
+    examples.main(name=None, force=False, branch='some-new-branch')
 
     mock_run.assert_called_once()
     captured = capsys.readouterr()
@@ -191,7 +276,18 @@ def test_force_clone(clone_examples, monkeypatch):
 
 def test_copy_example(clone_examples, tmp_directory):
     examples.main(name='templates/ml-online', force=False)
-    assert Path(tmp_directory, 'templates/ml-online').exists()
+
+    assert Path(tmp_directory, 'templates/ml-online').is_dir()
+    assert Path(tmp_directory, 'templates/ml-online', 'src',
+                'ml_online').is_dir()
+
+
+@pytest.mark.parametrize('target', ['custom-dir', 'custom/dir'])
+def test_copy_to_custom_directory(clone_examples, tmp_directory, target):
+    examples.main('templates/ml-online', output=target)
+
+    assert Path(tmp_directory, target).is_dir()
+    assert Path(tmp_directory, target, 'src', 'ml_online').is_dir()
 
 
 def test_error_unknown_example(clone_examples, capsys):
@@ -228,3 +324,38 @@ def test_error_if_git_clone_fails(monkeypatch):
         'An error occurred when downloading '
         'examples. Verify git is installed and your internet connection. '
         "(Error message: 'message')")
+
+
+@pytest.mark.parametrize('md, expected', [
+    ['', None],
+    ['<!-- start header -->\n', None],
+    ['\n\n<!-- end header -->\n\n', 2],
+    ["""
+<!-- start header -->
+
+<!-- end header -->
+""", 3],
+])
+def test_find_header(md, expected):
+    assert examples._find_header(md) == expected
+
+
+@pytest.mark.parametrize('md, clean', [
+    ['', ''],
+    ['there is no header', 'there is no header'],
+    ['stuff\n<!-- end header -->\nthings', 'things'],
+    ['more\nstuff\n<!-- end header -->\n\nthings', '\nthings'],
+])
+def test_skip_header(md, clean):
+    assert examples._skip_header(md) == clean
+
+
+@pytest.mark.parametrize('md, clean', [
+    ['', ''],
+    ['there is no header', 'there is no header'],
+    ['stuff\n<!-- end header -->\nthings', 'things'],
+    ['<!-- start description -->\nthings', 'things'],
+    ['<!-- end description -->\nthings', 'things'],
+])
+def test_cleanup_markdown(md, clean):
+    assert examples._cleanup_markdown(md) == clean